Add `Base` class instead of extending `Client`
