+++ lib/jortt/client.rb
@@ -1,3 +1,4 @@
+require 'jortt/client/base'
 require 'jortt/client/customers'
 require 'jortt/client/invoices'
 require 'jortt/client/invoice'
@@ -65,21 +66,6 @@ module Jortt
     # @return [ Jortt::Client::Invoice ] entry to the invoice resource.
     def invoice(invoice_id)
       Jortt::Client::Invoice.new(self, invoice_id)
-    end
-
-  private
-
-    def with_valid_response(&block)
-      block.call
-    rescue RestClient::Exception => error
-      raise Jortt::Error.new(error.message, error.response)
-    end
-
-    def with_valid_json(&block)
-      with_valid_response do
-        response = block.call
-        JSON.parse(response.body)
-      end
     end
   end
 end
